pyproject.toml
[project]
name = "token-dispenser"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "eth-ape>=0.8.36",
    "ruff>=0.12.1",
]
